removed mr_solver_set_precond; removed preconditions from MR Solver environments; changed mr_solver command to remember refinements that have been proved
